% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmt.R
\name{cdmt}
\alias{cdmt}
\title{Change Detection by Multispectral Trends}
\usage{
cdmt(
  sr_data,
  si_data,
  out_path = NULL,
  sr = NULL,
  si = NULL,
  years,
  cng_dir,
  th_const = 1.2,
  noise_rm = TRUE,
  cores = 1
)
}
\arguments{
\item{sr_data}{character. Yearly surface reflectance data. Either the name of a \code{SpatRaster} object in the global environment or the full path where rasters (in \emph{*.tif} format) are stored. Each layer name should include the year preceded by an underscore. If \code{NULL} only spectral indices are used.}

\item{si_data}{character. Yearly spectral indices. Either the name of a \code{SpatRaster} object in the global environment or the full path where rasters (in \emph{*.tif} format) are stored. Each layer name should include the year preceded by an underscore. If \code{NULL} only reflectance bands are used.}

\item{out_path}{character. The path where output rasters (in \emph{*.tif} format) are saved. Folders are created recursively if they do not exist. If \code{NULL} the output is a \code{SpatRaster} object created by the \pkg{terra} package.}

\item{sr}{numeric vector. Indices (positive integers) of the reflectance bands in each multiband raster to be included in the time series. If \code{NULL} all layers are used. Ignored if \code{sr_data} is a SpatRaster object.}

\item{si}{numeric vector. Indices (positive integers) of the spectral indices in each multiband raster to be included in the time series. If \code{NULL} all layers are used.}

\item{years}{numeric vector. Time interval (years) to be analysed.}

\item{cng_dir}{numeric vector. Direction of change caused by a disturbance in each spectral band/index. Valid values are either 1 or -1.}

\item{th_const}{numeric. Constant controlling the change threshold employed by the HiTS procedure during thresholding. Typical values are comprised in the interval \eqn{[1, 1.4]}.}

\item{noise_rm}{logical. If \code{TRUE} the impulsive noise filter is active.}

\item{cores}{positive integer. Number of CPU cores used during analysis.}
}
\value{
A \code{SpatRaster} object containing the following layers.
\item{EST}{Estimated values (one layer for each year and band).}
\item{CPT}{Detected changepoints (one layer for each year and band).}
\item{SLO}{Slope of the linear segments (one layer for each year and band).}
\item{MAG}{Magnitude in absolute terms (one layer for each year and band).}
\item{MAG_REL}{Magnitude in relative terms (one layer for each year and band).}
\item{LEN}{Length of segments (one layer per year).}
\item{CPT_ID}{Type of change (one layer per year). One of the following values: 101 (abrupt disturbance); 102 (abrupt greening); 201 (gradual disturbance); 202 (gradual greening); 9 (other change).}
\item{TPA}{Impulsive noise (one layer per year).}
\item{D_DUR}{Duration of disturbance (one layer per year).}
\item{G_DUR}{Duration of greening (one layer per year).}
\item{D_MAX}{Maximum disturbance change magnitude (in relative tems) throughout the time series (one layer per band).}
\item{D_FST}{Change magnitude (in relative tems) associated with the first disturbance detected within the time series (one layer per band).}
\item{G_MAX}{Maximum greening change magnitude (in relative tems) throughout the time series (one layer per band).}
\item{WGTS}{Weights assigned to each band (one layer per band).}
\item{D_MAX_MD}{Median among bands using values of \code{D_MAX} (single layer).}
\item{D_FST_MD}{Median among bands using values of \code{D_FST} (single layer).}
\item{G_MAX_MD}{Median among bands using values of \code{G_MAX} (single layer).}
\item{D_MAX_YR}{Year corresponding to \code{D_MAX_MD} (single layer).}
\item{D_FST_YR}{Year corresponding to \code{D_FST_MD} (single layer).}
\item{G_MAX_YR}{Year corresponding to \code{G_MAX_MD} (single layer).}
\item{D_MAX_DR}{Duration of the disturbance corresponding to \code{D_MAX_MD} (single layer).}
\item{D_FST_DR}{Duration of the disturbance corresponding to \code{D_FST_MD} (single layer).}
\item{G_MAX_DR}{Duration of the greening corresponding to \code{G_MAX_MD} (single layer).}
\item{D_MAX_ID}{Type of change (\code{CPT_ID}) of the disturbance corresponding to \code{D_MAX_MD} (single layer).}
\item{D_FST_ID}{Type of change (\code{CPT_ID}) of the disturbance corresponding to \code{D_FST_MD} (single layer).}
\item{G_MAX_ID}{Type of change (\code{CPT_ID}) of the greening corresponding to \code{G_MAX_MD} (single layer).}
\item{N_GAP}{Number of gaps in the time series, if any (single layer).}
\item{N_TPA}{Number of years containing impulsive noise, if any (single layer).}
\item{N_ITER}{Number of iterations performed by the impulsive noise filter (single layer).}
If a valid \code{out_path} is provided, a series of rasters in \emph{*.tif} format containing either one or multiple layers are directly written in the output folder.
}
\description{
This is the main function of the package.
}
\details{
\code{cdmt()} detects changes in linear trends of inter-annual Landsat time series at the pixel level.
Time series can be either univariate or multivariate, \emph{i.e.} include multiple spectral bands/indices.
Changes in the intercept, slope or both of linear trends are detected using the High-dimensional Trend Segmentation (HiTS) procedure proposed by Maeng (2019).
The HiTS procedure aims at detecting changepoints in a piecewise linear signal where the number and location of changepoints are unknown.
Input and output raster data is handled using the \pkg{terra} package.
}
\author{
Donato Morresi, \email{donato.morresi@gmail.com}
}
