% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmt_int.R
\name{cdmt_int}
\alias{cdmt_int}
\title{Pixel-wise Change Detection by Multispectral Trends}
\usage{
cdmt_int(
  x,
  nb,
  ny,
  yrs,
  ev = NA,
  dir,
  th_const = 1.2,
  noise_rm = TRUE,
  as_list = FALSE
)
}
\arguments{
\item{x}{numeric vector or matrix. If \code{x} is a numeric vector, it should contain sequences of yearly values for each band. If \code{x} is a matrix, each row should contain time-ordered data relative to an individual band.}

\item{nb}{integer. The number of bands in the time series.}

\item{ny}{integer. The number of years.}

\item{yrs}{numeric vector. The sequence of years to be analysed.}

\item{ev}{logical vector. A vector containing \code{NA} values to be used in case of processing failure.}

\item{dir}{numeric vector. Direction of change caused by a disturbance in each band. Valid values are either 1 or -1.}

\item{th_const}{numeric. Constant controlling the change threshold employed by the HiTS procedure during thresholding. Typical values are comprised in the interval \eqn{[1, 1.4]}.}

\item{noise_rm}{logical. If \code{TRUE} the impulsive noise filter is used.}

\item{as_list}{logical. If \code{TRUE} the output is a \code{list}. Otherwise, the output is a \code{numeric} \code{vector}.}
}
\value{
If \code{as_list} is \code{TRUE}, a \code{list} containing the following elements.
\item{est}{Estimated values (one layer for each year and band).}
\item{cpt}{Detected changepoints (one layer for each year and band).}
\item{slo}{Slope of the linear segments (one layer for each year and band).}
\item{mag}{Magnitude in absolute terms (one layer for each year and band).}
\item{mag_rel}{Magnitude in relative terms (one layer for each year and band).}
\item{len}{Length of segments (one layer per year).}
\item{cpt_id}{Type of change (one layer per year). One of the following values: 101 (abrupt disturbance); 102 (abrupt greening); 201 (gradual disturbance); 202 (gradual greening); 9 (other change).}
\item{tpa}{Impulsive noise (one layer per year).}
\item{d_dur}{Duration of disturbance (one layer per year).}
\item{g_dur}{Duration of greening (one layer per year).}
\item{d_max}{Maximum disturbance change magnitude (in relative tems) throughout the time series (one layer per band).}
\item{d_fst}{Change magnitude (in relative tems) associated with the first disturbance detected within the time series (one layer per band).}
\item{g_max}{Maximum greening change magnitude (in relative tems) throughout the time series (one layer per band).}
\item{wgts}{Weights assigned to each band (one layer per band).}
\item{d_max_md}{Median among bands using values of \code{D_MAX} (single layer).}
\item{d_fst_md}{Median among bands using values of \code{D_FST} (single layer).}
\item{g_max_md}{Median among bands using values of \code{G_MAX} (single layer).}
\item{d_max_yr}{Year corresponding to \code{D_MAX_MD} (single layer).}
\item{d_fst_yr}{Year corresponding to \code{D_FST_MD} (single layer).}
\item{g_max_yr}{Year corresponding to \code{G_MAX_MD} (single layer).}
\item{d_max_dr}{Duration of the disturbance corresponding to \code{D_MAX_MD} (single layer).}
\item{d_fst_dr}{Duration of the disturbance corresponding to \code{D_FST_MD} (single layer).}
\item{g_max_dr}{Duration of the greening corresponding to \code{G_MAX_MD} (single layer).}
\item{d_max-id}{Type of change (\code{CPT_ID}) of the disturbance corresponding to \code{D_MAX_MD} (single layer).}
\item{d_fst_id}{Type of change (\code{CPT_ID}) of the disturbance corresponding to \code{D_FST_MD} (single layer).}
\item{g_max_id}{Type of change (\code{CPT_ID}) of the greening corresponding to \code{G_MAX_MD} (single layer).}
\item{n_gap}{Number of gaps in the time series, if any (single layer).}
\item{n_tpa}{Number of years containing impulsive noise, if any (single layer).}
\item{n_iter}{Number of iterations performed by the impulsive noise filter (single layer).}
If \code{as_list} is \code{FALSE} the result is a \code{numeric} \code{vector}.
}
\description{
This is the internal function called by \code{\link{cdmt}} to operate at the pixel level.
}
\details{
\code{cdmt_int()} is not directly called by the user.
It analyses inter-annual Landsat time series to detect changes in spectral trends at the pixel level.
Time series can include single or multiple spectral bands/indices, hereafter referred to as bands.
Input data can be a \code{numeric} \code{vector} extracted from a \code{SpatRaster} or a \code{matrix}.
Impulsive noise, i.e. outliers in the time series, can be removed through an iterative procedure by the relevant filter.
One-year gaps in the time series are filled using either linear interpolation or extrapolation.
Changes in the intercept, slope or both of linear trends are detected using the High-dimensional Trend Segmentation (HiTS) procedure proposed by \insertCite{maeng2019adaptive;textual}{cdmt}.
The HiTS procedure aims at detecting changepoints in a piecewise linear signal where their number and location are unknown.
}
\examples{
library(cdmt)

# Load raster data
data(lnd_si)
lnd_si <- terra::rast(lnd_si)

# Extract values
v <- terra::values(lnd_si)[2000,]

# Process data
out <- cdmt_int(v,
                nb = 3,
                ny = 36,
                yrs = 1985:2020,
                dir = c(-1, 1, 1),
                as_list = TRUE
                )

# Plot results
bands <- c("MSI", "TCW", "TCA")
par(mfrow=c(length(bands), 1))

for(i in seq_along(bands)) {
  plot(matrix(v, nrow = 3, ncol = 36)[i,], type = "l", col = 2,
  lwd = 2, ylab = "", xlab = "", xaxt = "n")
  lines(out$est[i,], type = "l", col = 4, lwd = 2, lty = 1)
  abline(v = which(out$cpt_id == 101), lty = 2, col = 1, lwd = 2)
  axis(1, at = seq_along(1985:2020), labels = 1985:2020)
  title(main = paste(bands[i]), adj = 0)
}

}
\references{
\insertRef{maeng2019adaptive}{cdmt}
}
\author{
Donato Morresi, \email{donato.morresi@gmail.com}
}
